requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.9
Pillow>=10.3.0
celery==5.3.4
redis==5.0.1
django-redis==5.3.0
django-filter==23.3
drf-yasg==1.21.7
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cleanup==8.0.0
openpyxl==3.1.2
reportlab==4.0.6
python-multipart==0.0.6
dj-database-url==2.0.0
django-environ==0.10.0
requests>=2.31.0
# OCR识别相关依赖（发票智能识别功能）
paddleocr>=2.7.0
paddlepaddle>=3.0.0  # PaddleOCR的依赖
pdf2image>=1.16.3
pytesseract>=0.3.10
pypinyin>=0.49.0  # 中文拼音转换，用于菜单排序
